Change update method parameter to type double

+++ src/gamefield/GameField.java
@@ -52,7 +52,7 @@ public class GameField extends JPanel {
     }
 
 
-    public void update(long duration) {
+    public void update(double duration) {
         for (GameObject object : gameObjects) {
             object.update(duration);
         }
