+++ src/gameobjects/movable/car/EnemyCar.java
@@ -15,7 +15,7 @@ public class EnemyCar extends Car {
     }
 
     @Override
-    public void update(long duration) {
+    public void update(double duration) {
 
     }
 }
